fix: deploy scripts
