+++ packages/deploy/package.json
@@ -23,7 +23,7 @@
     "@sandbox-smart-contracts/dependency-royalty-management": "1.0.2",
     "@sandbox-smart-contracts/faucets": "0.0.1",
     "@sandbox-smart-contracts/giveaway": "0.0.3",
-    "@sandbox-smart-contracts/land": "*",
+    "@sandbox-smart-contracts/land": "0.0.3",
     "@sandbox-smart-contracts/marketplace": "1.0.1"
   },
   "files": [
